fastapi/requirements.txt
fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.22.0
pymongo==4.3.3
python-dotenv==1.0.0
motor==3.1.2